Add files via upload
This week due to a request by the professor, I finished the front design of the extension. UI design has been made by using Figma- a cloud-based design tool similar to Sketch, but with big differences that make Figma better for team collaboration. Colors have been chosen due to the trend of the year.  I have designed the logo, this year's trend is to hide the meaning behind the icons. By using the below link you can review the design through Figma. 
https://www.figma.com/file/6pRTrNiUb56BV8mxRO5S3k/Clivew-Extension